Single build and push
Optimise the tags that are pushed and can be used. As we only 
might need to use the latest, branch or commit sha these are the 
only ones that we are going to tag. Linked to this, as we don't need 
the branch tag on main, this stops it from pushing the branch_main 
tag unnecessarily.