HOCS-4537 Add helper scripts for connecting to db
This commit adds a few scripts that read the correct credentials from
the environment variables and feeds it into psql for a developer to use
to connect to one of the databases. At the moment `psql` defaults to a
default user that doesn't have any access to do anything.

Further work in this area could create some equivalent scripts for
read-only access. A cheap way to do that would be to set "SET SESSION
CHARACTERISTICS AS TRANSACTION READ ONLY;" in ~/.pgsqlrc.