Auto-deploy to environments on push
This commit changes behaviour of Drone to automatically deploy main to
all environments instead of waiting for a manual promotion.

It also removes pushing off-main versions to Quay, because there should
be no reason to not use mainline (and doing so would get overwritten
automatically, now).

This commit uses Anchor syntax to avoid repeating namespaces a lot.