HOCS-3069 Have "safe" and "dangerous" images
We currently have two images for 'toolbox' scripts, one for use in
production and most notprod environments, and another with some scary
scripts (like "drop the entire database") for a subset of notprod
environments.

The productionish image is poorly maintained and isn't very useful. This
commit modifies the Dockerfile of hocs-toolbox to create two images,
both based on the same base image. One contains the scary scripts, and
the other does not.

We also update the Drone file for this repository to push changes to
main to every environment simultaneously.