Autodeploy "safe" and "dangerous" versions of the toolbox, everywhere (#19)
* Auto-deploy to environments on push

This commit changes behaviour of Drone to automatically deploy main to
all environments instead of waiting for a manual promotion.

It also removes pushing off-main versions to Quay, because there should
be no reason to not use mainline (and doing so would get overwritten
automatically, now).

This commit uses Anchor syntax to avoid repeating namespaces a lot.

* HOCS-3069 Have "safe" and "dangerous" images

We currently have two images for 'toolbox' scripts, one for use in
production and most notprod environments, and another with some scary
scripts (like "drop the entire database") for a subset of notprod
environments.

The productionish image is poorly maintained and isn't very useful. This
commit modifies the Dockerfile of hocs-toolbox to create two images,
both based on the same base image. One contains the scary scripts, and
the other does not.

We also update the Drone file for this repository to push changes to
main to every environment simultaneously.

* Remove bare commit SHA version

* Use Postgres 14.2

* WIP: fix version ref

* WIP: fix syntax error

* WIP: comment out main filter for now

* WIP: install libpq

* WIP: add apk upgrade

* WIP: Move database scripts into safe

Previously the database scripts were in a directory that wasn't copied
to the image. This fixes that.

* HOCS-3069: change build step depends_on

Change the `build & push scary` depends_on to depend on the safe image
build. This is because of an issue with drone and the docker plugin.

Co-authored-by: Liam MacPherson <liammacpherson1996@gmail.com>