HOCS-5061 Fix database secrets
The database secrets in environments have moved. This commit fixes the
references to them, and exposes the new read-only credentials for use in
our scripts.

To save on lines of repeated YAML, we use the "split" feature of kd's Go
templating to define the secret structure only once.

This commit amends existing database scripts to use the RO version, and
adds extra read-write versions of the database scripts that will only
appear should "dangerous" be targeted in the Dockerfile.