Drop root after custom apk instead of before
You need root to run `apk install`, but we were dropping to a
non-privileged user before the custom apk steps.

This commit moves the drop to after the `apk install`, so that it
actually builds(!).

This has a downside of running `curl` as root which isn't ideal, but the
alternative is a complex Dockerfile or large layers.