Bring cms-migration branch up to main (#33)
* HOCS-5066 Fix db scripts not being in container (#23)

The previous COPY syntax copied only shell scripts. This commit changes
it to copy over whole folders, meaning that the db scripts are actually
copied into the image.

* Re-add accidentally uncommented lines

I commented out the "only on main" part of the pipeline while working on
my branch, but we forgot to add it back in before merging to main.

* Update README.md (#2)

* HOCS-5061 Fix database secrets

The database secrets in environments have moved. This commit fixes the
references to them, and exposes the new read-only credentials for use in
our scripts.

To save on lines of repeated YAML, we use the "split" feature of kd's Go
templating to define the secret structure only once.

This commit amends existing database scripts to use the RO version, and
adds extra read-write versions of the database scripts that will only
appear should "dangerous" be targeted in the Dockerfile.

* Mark database secrets as optional

This commit makes the database secrets optional. They're not actually
required to make the container useful, so it's okay to let the container
start without them.

* HOCS-5061 use correct database name secret and add casework database

* Remove accidental double-underscore

* HOCS-5133: Remove unwanted scripts

* HOCS-5133: Merge the dangerous and safe deployments but put the scripts in a different folder

* HOCS-5133: Replace Dockerfile with lightweight hocs base image

* HOCS-5133: Remove less util

* HOCS-5133: Rename deploySafe to deploy in drone.yml

* HOCS-5133: Remove kubectl from apk add

* HOCS-5133: Make cs-dev the default and remove from template

* HOCS-5133: fix shellcheck issue

* HOCS-5133: Add dronetree and tag events

* HOCS-5133: Add semver actions

* HOCS-5133: rename root to uid 0 from PR feedback

* HOCS-5133: Keep dockerfile running

* HOCS-5133: Keep dockerfile running

* HOCS-5061 Feedback from PR, remove redundant trigger

* HOCS-5061 Feedback from PR, deploy to wcs and cs prod

* HOCS-5061 Feedback from PR, change deployment to include prod

* HOCS-5061 Feedback from PR, add bash

* HOCS-5061 Feedback from PR, change deployment to include prod

* Revert "HOCS-5061 Feedback from PR, change deployment to include prod"

This reverts commit a587e1bf93d41e670ed38c0f83c4df79435de79b.

* Update kube/.drone.yml

Co-authored-by: Douglas Gardner <douglas@chippy.ch>

* HOCS-5061 Feedback from PR, merge scripts, change to downscaler

* HOCS-5061 add notprod and prod tokens

* HOCS-5061 remove reference to non-existent step

* Make script actually build

New-lines meant that the line-continuation characters didn't continue on
the lines they were meant to, meaning that the build failed because
`CURL` is not a valid Dockerfile command.

This commit removes the new-lines. For legibility we also move the
custom APK package section into a new Docker layer.

* Pull bucket name from secret

The bucket name is different in development and production, so we
can't hardcode it in the deployment.yaml script.

* Drop root after custom apk instead of before

You need root to run `apk install`, but we were dropping to a
non-privileged user before the custom apk steps.

This commit moves the drop to after the `apk install`, so that it
actually builds(!).

This has a downside of running `curl` as root which isn't ideal, but the
alternative is a complex Dockerfile or large layers.

Co-authored-by: Nick McMahon <nick.mcmahon@digital.homeoffice.gov.uk>