Fix secrets; add prod build step
This commit changes the CMS-specific dronefile to use CMS-specific
Kubernetes tokens that don't have access to any other environment, so we
can't accidentally deploy the wrong version somewhere.

As this is a manual build I've also set it to autodeploy to production
on push once the deployment to dev succeeds. Drone's manual promote
feature doesn't work well with multiple branches, and I think it's
probably easier for everyone involved if hocs-toolbox is the same
version in both dev and prod at all times.