HOCS-5050 Add script to simplify case unsticking
A common user support request is to unstick a case that is unable to be
progressed due to the problem introduced during a recent P2 (HOCS-5050;
INC2383511) incident. At time of commit we have had 84 of these
requests, each usually containing multiple cases, which is a significant
drain on resources.

The process of dealing with one of these requests is usually:
  1. Find the article on Confluence that explains how to do it
  2. Copy a complex SQL command into a text editor
  3. Replace placeholders with real references
  4. Paste the SQL commands into a database REPL
  5. Go back to step 2 and use the database output to fill in
     the placeholders of the next command

This is time-consuming and errors can be easily introduced.

This commit adds a script that runs all those SQL commands for you,
making responding to one of these tickets much quicker. It echoes the
results to standard output, so that a developer can spend more time
figuring out the correct rectifying `UPDATE` commands without having to
worry about the preliminary `SELECT` commands.

Additionally to this, the script does some basic calculations to look up
what is probably the correct `UPDATE` command for the most common
rectification scenarios, and outputs them to stdout, ready for a human
to double-check them before running them on the live database.

The tool depends on the db scripts that are in the `scripts/db`
directory, but only uses the read-only scripts.